03-niche-finetuned-model/requirements.txt
# Unsloth and transformers
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers==4.37.2
torch==2.1.2
xformers==0.0.23.post1

# PEFT and quantization
peft==0.8.2
bitsandbytes==0.42.0
accelerate==0.26.1

# Training utilities
datasets==2.16.1
trl==0.7.10
wandb==0.16.2

# Data curation
beautifulsoup4==4.12.3
requests==2.31.0
html2text==2024.2.26

# UI and utilities
streamlit==1.31.0
python-dotenv==1.0.0
tqdm==4.66.1

# Evaluation
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# Hugging Face
huggingface-hub==0.20.3
